Return XTZ specified functions back to xtz/util.ts and put hashTypes back to xtz/utils.ts
Related unit tests refined
